Add validation for nested passenger, failing to book REDIRECTS to root
Instead of rendering a new bookings form we redirect because when we
render booking#new, it has @flight used
and @flight needs params['flight_id'] to be found.
Because we render new, the params doesn't PERSIST and @flight becomes nil
and new has @flight.departure_airport etc. methods which can't work
on nil giving error. Hence I came up with solution of redirecting to
root page with error msg.

IDK how viable this is and how hacky this is